# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentAny - Peer-to-Peer Rental Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products with optional filters
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProductById
      tags: [products]
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product listing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/owner/{ownerId}:
    get:
      operationId: getProductsByOwner
      tags: [products]
      summary: Get products by owner ID
      parameters:
        - in: path
          name: ownerId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner's products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/product/{productId}:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewDetail"
...
  /admin/products:
    get:
      operationId: getAdminProducts
      tags: [admin]
      summary: Get all products (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, owner, admin]
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        rating:
          type: number
        trustScore:
          type: number
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - rating
        - trustScore
        - isBlocked
        - createdAt
    UpdateUserInput:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        location:
          type: string
        phone:
          type: string
        avatar:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        pricePerDay:
          type: number
        deposit:
          type: number
        location:
          type: string
        images:
          type: array
          items:
            type: string
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatar:
          type: string
          nullable: true
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - pricePerDay
        - deposit
        - location
        - images
        - ownerId
        - ownerName
        - rating
        - reviewCount
        - isAvailable
        - createdAt
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            owner:
              $ref: "#/components/schemas/User"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/ReviewDetail"
            unavailableDates:
              type: array
              items:
                type: string
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
    CreateProductInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        pricePerDay:
          type: number
        deposit:
          type: number
        location:
          type: string
        images:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - category
        - pricePerDay
        - deposit
        - location
    UpdateProductInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        pricePerDay:
          type: number
        deposit:
          type: number
        location:
          type: string
        images:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - productId
        - bookingId
        - rating
        - comment
        - createdAt
    ReviewDetail:
      allOf:
        - $ref: "#/components/schemas/Review"
        - type: object
          properties:
            userName:
              type: string
            userAvatar:
              type: string
              nullable: true
    CreateReviewInput:
      type: object
      properties:
        productId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - productId
        - bookingId
        - rating
        - comment
...
    AdminProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...